+++ CHANGES.rst
@@ -2,12 +2,12 @@ Changelog
 =========
 
 1.2.0 (upcoming)
-----------------
+------------------
 
 New Features
 ~~~~~~~~~~~~
 - Now supports Python 3.10 [#19]
-
+- HOD module now works with halo light cone catalogs [#28]
 
 1.1.0 (2022-01-21)
 ------------------